created exceptions

+++ src/main/java/io/github/leonarddepaula/libaryapi/service/AutorService.java
@@ -1,7 +1,11 @@
 package io.github.leonarddepaula.libaryapi.service;
 
+import io.github.leonarddepaula.libaryapi.exceptions.OperacaoNaoPermitida;
 import io.github.leonarddepaula.libaryapi.model.Autor;
 import io.github.leonarddepaula.libaryapi.repository.AutorRepository;
+import io.github.leonarddepaula.libaryapi.repository.LivroRepositoty;
+import io.github.leonarddepaula.libaryapi.validador.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,16 +13,18 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor // essa anotation cria um contrutor em tempo de compilaçao.
 public class AutorService {
 
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepositoty livroRepository;
 
-    public AutorService(AutorRepository repository) {
-        this.repository = repository;
-    }
 
     public Autor salvar(Autor autor) {
+
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -26,6 +32,8 @@ public class AutorService {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("Para Atulizar é nescessário que já exista um autor");
         }
+
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -35,6 +43,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitida("Autor possui Livros Cadastrados ** EXCLUSSÃO NÃO PERMITIDA");
+        }
         repository.delete(autor);
     }
 
@@ -52,5 +64,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
